chore: fix linting errors
